Refactor: 피드백 반영하여 코드 전반 리팩토링:
