+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/ice_breaking/controller/GameController.java
@@ -2,20 +2,30 @@ package com.garamgaebi.GaramgaebiServer.domain.ice_breaking.controller;
 
 import com.garamgaebi.GaramgaebiServer.domain.ice_breaking.entity.ProgramGameroom;
 import com.garamgaebi.GaramgaebiServer.domain.ice_breaking.dto.*;
-import com.garamgaebi.GaramgaebiServer.domain.ice_breaking.service.GameService;
+import com.garamgaebi.GaramgaebiServer.domain.ice_breaking.service.GameServiceImpl;
 import com.garamgaebi.GaramgaebiServer.global.response.BaseResponse;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@Tag(name = "Ice_breaking Controller", description = "아이스 브레이킹 컨트롤러(담당자: 애플)")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/game")
 public class GameController {
-    private final GameService gameService;
+    private final GameServiceImpl gameService;
 
     // programIdx로 게임방 조회
+    @Operation(summary = "게임방 리스트 조회", description = "programIdx로 해당하는 네트워킹의 게임방을 조회", responses = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "404", description = "존재하지 않는 리소스", content = @Content()),
+            @ApiResponse(responseCode = "500", description = "알 수 없는 서버 에러", content = @Content())
+    })
     @GetMapping("/{programIdx}/rooms")
     @ResponseBody
     public BaseResponse<List<ProgramGameroom>> getRooms(@PathVariable Long programIdx) {
@@ -23,6 +33,11 @@ public class GameController {
     }
 
     // gameRoomIdx로 members 조회
+    @Operation(summary = "멤버 리스트 조회", description = "gameRoomIdx로 해당하는 게임방의 멤버 리스트를 조회", responses = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "404", description = "존재하지 않는 리소스", content = @Content()),
+            @ApiResponse(responseCode = "500", description = "알 수 없는 서버 에러", content = @Content())
+    })
     @PostMapping("/members")
     @ResponseBody
     public BaseResponse<List<MembersGetRes>> getMembers(@RequestBody MembersGetReq membersGetReq) {
@@ -30,6 +45,11 @@ public class GameController {
     }
 
     // gameRoom에 member 등록
+    @Operation(summary = "멤버 등록", description = "roomId로 멤버를 게임방에 등록", responses = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "404", description = "존재하지 않는 리소스", content = @Content()),
+            @ApiResponse(responseCode = "500", description = "알 수 없는 서버 에러", content = @Content())
+    })
     @PostMapping("/member")
     @ResponseBody
     public BaseResponse<MemberRoomRes> postMemberToGameRoom(@RequestBody MemberRoomReq memberRoomReq) {
@@ -37,13 +57,23 @@ public class GameController {
     }
 
     // gameRoom에 member 삭제
+    @Operation(summary = "멤버 삭제", description = "roomId로 멤버를 게임방에서 삭제", responses = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "404", description = "존재하지 않는 리소스", content = @Content()),
+            @ApiResponse(responseCode = "500", description = "알 수 없는 서버 에러", content = @Content())
+    })
     @DeleteMapping("/member")
     @ResponseBody
-    public BaseResponse<MemberRoomRes> deleteMemberFromGameRoom(@RequestBody MemberRoomReq memberRoomReq) {
+    public BaseResponse<String> deleteMemberFromGameRoom(@RequestBody MemberRoomReq memberRoomReq) {
         return new BaseResponse<>(gameService.deleteMemberFromGameRoom(memberRoomReq));
     }
 
     // 랜덤 이미지 조회
+    @Operation(summary = "랜덤 이미지 30장 조회", description = "programIdx를 seed로 하여 랜덤 이미지 30장 조회", responses = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "404", description = "존재하지 않는 리소스", content = @Content()),
+            @ApiResponse(responseCode = "500", description = "알 수 없는 서버 에러", content = @Content())
+    })
     @GetMapping("/{programIdx}/images")
     @ResponseBody
     public BaseResponse<List<String>> getGameImages(@PathVariable Long programIdx) {
@@ -51,6 +81,11 @@ public class GameController {
     }
 
     // 게임방 현재 이미지 인덱스 증가
+    @Operation(summary = "게임방 이미지 인덱스 증가", description = "현재 게임방의 current image index 증가", responses = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "404", description = "존재하지 않는 리소스", content = @Content()),
+            @ApiResponse(responseCode = "500", description = "알 수 없는 서버 에러", content = @Content())
+    })
     @PatchMapping("/current-idx")
     @ResponseBody
     public BaseResponse<String> patchCurrentImgIdx(@RequestBody CurrentImgIdxReq currentImgIdxReq) {
