+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/ice_breaking/dto/CurrentImgIdxReq.java
@@ -1,8 +1,9 @@
 package com.garamgaebi.GaramgaebiServer.domain.ice_breaking.dto;
 
-import lombok.Data;
+import lombok.*;
 
-@Data
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class CurrentImgIdxReq {
     private String roomId;
 }
