+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/ice_breaking/dto/MemberRoomRes.java
@@ -3,10 +3,14 @@ package com.garamgaebi.GaramgaebiServer.domain.ice_breaking.dto;
 import lombok.*;
 
 @Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PUBLIC)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class MemberRoomRes {
     private String message;
     private int currentImgIdx;
+
+    @Builder
+    public MemberRoomRes(String message, int currentImgIdx) {
+        this.message = message;
+        this.currentImgIdx = currentImgIdx;
+    }
 }
