+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/ice_breaking/dto/MembersGetRes.java
@@ -4,12 +4,17 @@ package com.garamgaebi.GaramgaebiServer.domain.ice_breaking.dto;
 import lombok.*;
 
 @Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PUBLIC)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class MembersGetRes {
     private Long memberIdx;
     private String nickname;
 
     private String profileUrl;
+
+    @Builder
+    public MembersGetRes(Long memberIdx, String nickname, String profileUrl) {
+        this.memberIdx = memberIdx;
+        this.nickname = nickname;
+        this.profileUrl = profileUrl;
+    }
 }
