+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/ice_breaking/dto/MessageDto.java
@@ -1,14 +1,9 @@
 package com.garamgaebi.GaramgaebiServer.domain.ice_breaking.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 @Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class MessageDto {
     public enum MessageType {
         ENTER, TALK, EXIT
@@ -19,4 +14,13 @@ public class MessageDto {
     private String sender; // 보내는 사람
     private String message; // 사진 index
     private String profileUrl; // profile image url
+
+    @Builder
+    public MessageDto(MessageType type, String roomId, String sender, String message, String profileUrl) {
+        this.type = type;
+        this.roomId = roomId;
+        this.sender = sender;
+        this.message = message;
+        this.profileUrl = profileUrl;
+    }
 }
