+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/member/dto/EmailRes.java
@@ -3,9 +3,12 @@ package com.garamgaebi.GaramgaebiServer.domain.member.dto;
 import lombok.*;
 
 @Getter
-@Setter
-@AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PUBLIC)
 public class EmailRes {
     private String message;
+
+    @Builder
+    public EmailRes(String message) {
+        this.message = message;
+    }
 }
