+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/member/dto/InactivedMemberRes.java
@@ -3,9 +3,12 @@ package com.garamgaebi.GaramgaebiServer.domain.member.dto;
 import lombok.*;
 
 @Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PUBLIC)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class InactivedMemberRes {
     private boolean inactive_success;
+
+    @Builder
+    public InactivedMemberRes(boolean inactive_success) {
+        this.inactive_success = inactive_success;
+    }
 }
