+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/member/dto/MemberLogoutReq.java
@@ -1,10 +1,9 @@
 package com.garamgaebi.GaramgaebiServer.domain.member.dto;
 
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 
 @Getter
-@Setter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class MemberLogoutReq {
     private String accessToken;
     private String refreshToken;
