+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/member/dto/MemberRolesDto.java
@@ -4,11 +4,16 @@ import com.garamgaebi.GaramgaebiServer.domain.member.entity.MemberRoles;
 import lombok.*;
 
 @Getter
-@Setter
-@NoArgsConstructor(access = AccessLevel.PUBLIC)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class MemberRolesDto {
     private Long memberIdx;
     private String roles;
+
+    @Builder
+    public MemberRolesDto(Long memberIdx, String roles) {
+        this.memberIdx = memberIdx;
+        this.roles = roles;
+    }
 
     public MemberRoles toEntity() {
         return MemberRoles.builder()
