+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/member/dto/PostMemberReq.java
@@ -13,19 +13,6 @@ public class PostMemberReq {
     private String uniEmail;
     private MemberStatus status;
 
-    @Builder
-    public PostMemberReq(String nickname,
-                         String profileEmail,
-                         String socialEmail,
-                         String uniEmail,
-                         MemberStatus status) {
-        this.nickname = nickname;
-        this.profileEmail = profileEmail;
-        this.socialEmail = socialEmail;
-        this.uniEmail = uniEmail;
-        this.status = status;
-    }
-
     public Member toEntity() {
         return Member.builder()
                 .nickname(nickname)
