+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/member/dto/PostMemberRes.java
@@ -3,9 +3,12 @@ package com.garamgaebi.GaramgaebiServer.domain.member.dto;
 import lombok.*;
 
 @Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PUBLIC)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PostMemberRes {
     private Long memberIdx;
+
+    @Builder
+    public PostMemberRes(Long memberIdx) {
+        this.memberIdx = memberIdx;
+    }
 }
