+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/member/dto/VerifyEmailReq.java
@@ -1,8 +1,12 @@
 package com.garamgaebi.GaramgaebiServer.domain.member.dto;
 
+import lombok.AccessLevel;
 import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-@Data
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class VerifyEmailReq {
     private String email;
     private String key;
